# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/roles": {
      "post": {
        "operationId": "RolesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "get": {
        "operationId": "RolesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/roles/{id}": {
      "get": {
        "operationId": "RolesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "patch": {
        "operationId": "RolesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "delete": {
        "operationId": "RolesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/roles/name/{roleName}/staff": {
      "get": {
        "operationId": "RolesController_getStaffByRoleName",
        "parameters": [
          {
            "name": "roleName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Staff"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/roles/{id}/staff": {
      "get": {
        "operationId": "RolesController_getStaffByRoleId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Staff"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
...
    "/staff-register/{staffId}/roles/{roleId}": {
      "delete": {
        "operationId": "StaffRegisterController_removeRole",
        "parameters": [
          {
            "name": "staffId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "StaffRegister"
        ]
      }
    },
...
    "/notification/{id}": {
      "get": {
        "operationId": "NotificationController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Notification"
        ]
      },
      "patch": {
        "operationId": "NotificationController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "Notification"
        ]
      },
      "delete": {
        "operationId": "NotificationController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Notification"
        ]
      }
    },
...
    "/push-notification/subscribe": {
      "post": {
        "operationId": "PushNotificationController_subscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePushNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PushNotification"
        ]
      }
    },
...
    "/push-notification/{id}": {
      "get": {
        "operationId": "PushNotificationController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PushNotification"
        ]
      }
    },
...
    "/push-notification/unsubscribe/{id}": {
      "delete": {
        "operationId": "PushNotificationController_unsubscribe",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PushNotification"
        ]
      }
    },
...
    "/quote": {
      "post": {
        "operationId": "QuoteController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          }
        },
        "tags": [
          "Quote"
        ]
      },
      "get": {
        "operationId": "QuoteController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Quote"
        ]
      }
    },
...
    "/quote/{id}": {
      "get": {
        "operationId": "QuoteController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          }
        },
        "tags": [
          "Quote"
        ]
      },
      "patch": {
        "operationId": "QuoteController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          }
        },
        "tags": [
          "Quote"
        ]
      },
      "delete": {
        "operationId": "QuoteController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Quote"
        ]
      }
    },
...
  },
  "info": {
    "title": "CN Squad API",
    "description": "The CN Squad API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "permissions": {
            "minItems": 1,
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "permissions"
        ]
      },
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "action": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "action",
          "createdAt",
          "updatedAt"
        ]
      },
      "StaffAddress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "city",
          "state",
          "country",
          "createdAt",
          "updatedAt"
        ]
      },
      "Department": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nameAbrv": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nameAbrv",
          "name",
          "description",
          "createdAt",
          "updatedAt"
        ]
      },
      "DepartmentalRole": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "department",
          "createdAt",
          "updatedAt"
        ]
      },
      "StaffEmployment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "employeeCode": {
            "type": "string"
          },
          "jobTitle": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "employmentType": {
            "type": "string",
            "enum": [
              "Full-Time",
              "Part-Time",
              "Contract",
              "Intern"
            ]
          },
          "workMode": {
            "type": "string",
            "enum": [
              "Onsite",
              "Remote",
              "Hybrid"
            ]
          },
          "hireDate": {
            "type": "string"
          },
          "reportingManager": {
            "type": "string"
          },
          "directReport": {
            "type": "string"
          },
          "contractEndDate": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "Suspended",
              "Terminated"
            ]
          },
          "workLocation": {
            "type": "string"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "departmentalRole": {
            "$ref": "#/components/schemas/DepartmentalRole"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "employeeCode",
          "jobTitle",
          "employmentType",
          "workMode",
          "hireDate",
          "reportingManager",
          "directReport",
          "status",
          "department",
          "departmentalRole",
          "createdAt",
          "updatedAt"
        ]
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          },
          "staff": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Staff"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "permissions",
          "staff",
          "createdAt",
          "updatedAt"
        ]
      },
      "Staff": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "uuid": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "hobbies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dateOfBirth": {
            "type": "string"
          },
          "gender": {
            "enum": [
              "Male",
              "Female",
              "Other"
            ],
            "type": "string"
          },
          "maritalStatus": {
            "enum": [
              "Single",
              "Married",
              "Divorced",
              "Widowed"
            ],
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/StaffAddress"
          },
          "employment": {
            "$ref": "#/components/schemas/StaffEmployment"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "resetToken": {
            "type": "string"
          },
          "resetTokenExpires": {
            "format": "date-time",
            "type": "string"
          },
          "registrationToken": {
            "type": "string"
          },
          "registrationTokenExpires": {
            "format": "date-time",
            "type": "string"
          },
          "isRegistered": {
            "type": "boolean"
          },
          "lastIpAddress": {
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isCeo": {
            "type": "boolean"
          },
          "isOnline": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "uuid",
          "firstName",
          "lastName",
          "description",
          "hobbies",
          "dateOfBirth",
          "gender",
          "email",
          "phone",
          "password",
          "address",
          "employment",
          "roles",
          "isRegistered",
          "isCeo",
          "isOnline",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {}
      },
...
      "Admin": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "staff": {
            "$ref": "#/components/schemas/Staff"
          },
          "password": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "isAdmin",
          "firstName",
          "lastName",
          "email",
          "staff",
          "password",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateNotificationDto": {
        "type": "object",
        "properties": {}
      },
...
      "PushSubscriptionDataDto": {
        "type": "object",
        "properties": {
          "p256dh": {
            "type": "string"
          },
          "auth": {
            "type": "string"
          }
        },
        "required": [
          "p256dh",
          "auth"
        ]
      },
      "CreatePushNotificationDto": {
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string",
            "format": "uri"
          },
          "keys": {
            "$ref": "#/components/schemas/PushSubscriptionDataDto"
          }
        },
        "required": [
          "endpoint",
          "keys"
        ]
      },
...
      "Quote": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "fileUrl": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdBy": {
            "$ref": "#/components/schemas/Admin"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subject",
          "description",
          "fileUrl",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}